Fixed hidden bug in process macro leading to wrong iteration values in the loop body:
iteration values in for clauses were end-tested before and incremented
within the looping body of the process (consisting of the content of
'sprout, 'output or 'wait; called 'looping in cm's parser) just before
the actual looping code:

Before:

- the iteration variables got initialized in an outer
  'let. for... = were initialized with nil

- then all for...= intializations were done

- the closed over function of the process contained

  1. The end-test of the loop (all test-clauses combined with 'or).
  2. the stepping-function of variables without an end test
  3. The looping body
  4. The stepping-function of the variables with end-test

This created wrong behaviour in the following construct:

(process
  for i below 3
  for count = i then (+ count 1)
  output (format t "i: ~a, count: ~a~%")
  wait 0.5)

The count variable is initialized to 0 outside the function and then
incremented before the output routine is executed, leading to a count
value of 1 already in the first iteration.

The code has been changed to the following scheme (following in part
the way 'loop is expanded in sbcl):

- the iteration variables are initialized in an outer 'let.
  for... = are initialized with nil

- the values of iteration variables without end-test are initialized
  after the end-test is performed (outside the closed over function)

- the closed over function contains

  1. the looping body
  2. the stepping-function of the variables with end-test
  3. The end-test of the loop (all test-clauses combined with 'or)
  4. the stepping-function of variables without an end test

In order to achieve this, an additional 'stepping-notest was created,
as the previous version simply pushed these variable rebindings into
'looping.

NOTE: the cltl2-loop function also uses the modified iteration
functions so this is momentarily broken. As the bug also effects this
function, it should get corrected anyway and will be fixed soon...

The current commit still contains the old function.